refactor #31: login 관련 비지니스 로직 디렉토리 수정
