+++ src/app/(auth)/login/hooks/useLoginForm.ts
@@ -1,12 +1,22 @@
-import { UseMutateAsyncFunction } from '@tanstack/react-query';
+import { useRouter } from 'next/navigation';
+import { useMutation } from '@tanstack/react-query';
+import { login } from '@api/auth';
 import { ApiError } from '@api/types';
 import useForm from '@hooks/useForm';
-import { Login, LoginFormInput } from '@mutations/useLoginMutaion/types';
 
-interface useLogInFormProps {
-  mutateAsync: UseMutateAsyncFunction<Login, ApiError, LoginFormInput, unknown>;
-}
-const useLogInForm = ({ mutateAsync }: useLogInFormProps) => {
+const useLogInForm = () => {
+  const router = useRouter();
+
+  const mutation = useMutation(login, {
+    onSuccess: ({ token }) => {
+      localStorage.setItem('token', token);
+      router.push('/main');
+    },
+    onError: ({ reason }: ApiError) => {
+      alert(reason);
+    },
+  });
+
   const { values, handleChange, handleSubmit } = useForm({
     initialState: {
       username: '',
@@ -15,12 +25,15 @@ const useLogInForm = ({ mutateAsync }: useLogInFormProps) => {
     validate: ({ password, username }) => {
       let passwordError = null;
       let usernameError = null;
+
       if (!username) {
         passwordError = '이메일을 입력해주세요.';
       }
+
       if (!password) {
         usernameError = '비밀번호를 입력해주세요.';
       }
+
       return {
         password: passwordError,
         username: usernameError,
@@ -28,12 +41,15 @@ const useLogInForm = ({ mutateAsync }: useLogInFormProps) => {
     },
     onSubmit: async () => {
       const { username, password } = values;
-      await mutateAsync({ username, password });
+      await mutation.mutateAsync({ username, password });
     },
   });
+
   return {
-    handleSubmit,
+    ...mutation,
     handleChange,
+    handleSubmit,
   };
 };
+
 export default useLogInForm;
